Enhancement: Can not register array's of objects
You can now register an array of items by using a suffix on a name of
'[]'. When the name is suffixed with [] it won't store the value but
will instead add the item to an array which is stored instead. Any new
items will be pushed to the end of the same array.

Any registration with the same name but without the suffix will
de-register the array and vice versa.