Enhancements: Added dispose and registration class
`dispose` is a a place holder method for now. The class is now extending
event emitter and when `dispose()` is called it will emit a `preDispose`
and `postDispose` event - but it doesn't actually do anything else.

`registration` class in place. This is going to be used to store details
about each registration.